Refactored lint.js to improve readability
The logging of the Joker message was moved out to a separate function and called from the for loop. The length checks were unnecessary as if the data.toString() results in the empty string nothing will happen. 

Minor refactoring of spacing to be consistent with rest of the file. Lets that were never modified have been changed to consts.