Allow toggling Paredit function independently of mode
Closes #784

We introduce a `calva.paredit.keybindingsEnabled` config option and a
`paredit.toggleKeybindingsEnabled` command which allows toggling Paredit
keybindings on and off while preserving the current mode. The Paredit
keybindings are updated to only activate when the
`paredit:keybindingsEnabled` context key is true, in addition to the
existing conditions on the keymap. Finally, the Paredit StatusBar
component is updated to receive a `KeyMapConfig` object containing
`keyMap` and the new `keybindingsEnabled` boolean, and a new path is
added in the UI state management to handle this case as distinct from the
'none' keyMap case.