Improve #942 (slow repl appending) by concating pending output events
Basically does what klauswuestefeld suggested here:

https://github.com/BetterThanTomorrow/calva/issues/942#issuecomment-759049579

If there is a queue, can't we read all pending output events from the
queue, concat them and apply them as a single batch to the output file?
That would maybe make the console jumpy but at least it would not be
absurdly slow.