Factor all test functions in terms of Var Query
Since a set of recent changes, all of the test runs that Calva runs
are implemented using Cider's test-var-query operation. This allows
us to re-write test, testNs, and testAll all in terms of a new function
to testVarQuery.

We can also define a concrete type for the var-query (and ns-query)
in cider.ts to allow us to make changes more safely.