Create clojure-lsp drag commands
This is the start of exposing the clojure-lsp drag-backward and
drag-forward commands as targets for keybindings. This was discussed in
https://github.com/BetterThanTomorrow/calva/discussions/1695

These commands are similar to the built-in Calva commands
dragSexprForward and dragSexprBackward. But users may want to use the
clojure-lsp versions to take advantage of additional features that
clojure-lsp offers:

- Better comment affinity. Comments stay affiliated with code blocks as
  they are dragged.
- Better contextual support. Clauses within many common forms (e.g. case
  and cond) move intuitively.

Resolves #1697.