Disable default json formatter to avoid conflict with prettier
The default formatter removes spaces at the end of files, and prettier adds them. ESLint will complain if the newline is not there.

When using format on save, if you edit the package.json file, for example, and the default formatter is not enabled, the newline at the end of the file will be removed. When the prettier check is run by CI, it'll fail.