Handle multiple files in ProjectType.useWhenExists
Gradle supports both a Groovy DSL using "*.gradle" files and a Kotlin
DSL using "*.gradle.kts" files. To properly detect Gradle projects we
needed a way to detect any of these as an option.

Since this property was only used in the project-types.ts file, it
shouldn't be a breaking change for anyone.