Handle completion response failures gracefully
The provider completions subsystem requests results from the lsp and
nrepl providers sequentially and then merges the results together. If
either of these providers fail to report completion items and throws
an error then the entire flow fails.

Ideally if one of the providers is working and is able to report
results then the completions should continue to work.

This commit fixes this issue by catching errors from the completions
provider and quietly logging them to console instead of bubbling the
error upstream.

Fixes #2006