Add behaviours to `enableClojureLspOnStart` config
This commit updates the `enableClojureLspOnStart` setting to allow
specifying different auto-start behaviours. The setting now accepts the
values:

- "when-workspace-opened-use-workspace-root"
- "when-file-opened-use-furthest-project"
- "never"
- true
- false

And has slightly different behaviours in each case. The value `true` is
now an alias for `"when-workspace-opened-use-workspace-root"`, and
`false` for `"never"`.

When set to `"when-workspace-opened-use-workspace-root"` the clojure-lsp
will be automatically started in the root of all open/added workspaces
when calva initializes and whenever a workspace is added to vscode.

When set to `"when-file-opened-use-furthest-project"` then the
clojure-lsp will be automatically started whenever a clojure file is
opened. The server will be started in the furthest possible valid
clojure project or will fall back to starting in the workspace root if
no valid clojure project is found.

When set to `"never"` the clojure-lsp will never automatically be
started and is instead left up to the user to explicitly start it in a
directory.