Only start clojure-lsp in distinct workspace roots
Currently when working in Calva with the
"when-workspace-opened-use-workspace-root" setting Calva will attempt to
auto-start the clojure-lsp in _all_ discovered valid clojure projects.
This causes problems when both a projects root and subdirectories from
that project are opened simultaneously.

To resolve this issue we skip starting the clojure-lsp for workspace
folders which are decendents of already opened workspaces that have
running lsp servers.

Additionally files opened in descendent workspace folders were not able
to interact with lsp processes spawned for other workspace folders, even
if those folders were parent directories.

This is now resolved in the client selector which removes the artificial
traversal limit.

Fixes #2065