Fallback to starting lsp-server in tmp dir
Currently we do not auto-start the lsp server if no valid clojure
projects are found. This means that we won't have any lsp support for
files in workspace folders that do not contain project files. This can
be quite unexpected.

Additionally this means we don't support lsp operations on files opened
which are not part of any workspace (for example when peeking or using
File > Open on a random file on the host).

This commit makes two changes:

+ Adds a fallback behaviour of provisioning an lsp server in a OS tmp
  directory if no valid project is found.
+ Adds `.lsp/config.edn` as a heuristic for a valid clojure project.

Fixes #2069