Disable fallback clojure-lsp server
The clojure-lsp fallback server does not work as expected on Windows
machines.

The function to start the fallback server has now been changed to return
early until support for Windows can be safely added.

Additionally the logic for `"always-use-first-workspace-root"` now
always starts the lsp server in the first project root instead of
starting the fallback server if the first project root is not a valid
clojure project.

As much as possible the code for working with the fallback server has
been left along in order to facilitate re-enabling these code-paths once
Windows support has been figured out.

Closes #2088
Fixes #2087