Fix lsp client path comparisons on Windows
The api method for finding an LSP client for a given document URI was
written to stop traversing when encountering the OS root. This logic
assumed POSIX paths and did not work correctly on Windows platforms.

This updates the clojure-lsp client ids to use the uri.fsPath which is
platform agnostic and updates the root path comparison in the lsp api to
compare in a platform agnostic manner.