Use Uri.file instead of Uri.parse for file paths
The `Uri.parse` method is for parsing full URI's that include a 'scheme'
component. Our usage was 'fine' for POSIX systems as it was non-strict
and would fallback to a `file://` scheme however on Windows it would
interpret the leading `C:\` as the scheme.

This commit replaces all our usages of `parse` with `file` where we are
using known file paths to remove this ambiguity.