Automatically show rename UI after refactors that add names
Fixes #2388

Most language extensions automatically trigger the rename UI after
completing a refactor command that introduces new names, such as
Javascript's "Extract to constant" commands. Previously, for such
commands run from the command palette, Calva prompted for the new name
in a palette, and did not support auto-renaming commands run from
quick-fix actions (under the lightbulb icon), whatever default name was
returned from the LSP would be used. Now, Calva uses the Rename Symbol
feature in both cases.

To support this, introduce two new properties of ClojureLspCommand,
`defaultName` and `afterCommandFn`. `afterCommandFn` is called after the
initial command completes when it is present on the command object. The
only implementation of an `afterCommandFn` presently is
`renameAfterRefactor`, which looks for the hardcoded `defaultName` in
the edit returned from the LSP and triggers Rename Symbol on the last
place where it appears, leaving the user's cursor near where it started.

Also extract clojureLspCommands and moved it above sendCommandRequest
because sendCommandRequest uses it to retrieve the afterCommandFn from
the original command object, which is not otherwise present in scope.