Generate .circleci/config.yml from config.ys
This is the start of refactoring.

config.yml changed massively here, but should never change afterwards.

The goal is to make the input YAML/YS nice and clean and generate the
correct config.yaml.

Added a simple Makefile to do the generation.
It auto-installs the `ys` binary into /tmp if needed.
So you don't need to worry about that.

Run from the top level directory with:

    make -C .circleci/

or:

    make -C .circleci/ clean build

if you need to force it.