Format (ns) and remainder separately to balance multicursor & sort-ns-references
To nudge all cursors with the tides of reformatting, micro edits change only
whitespace. Computing edits is linear-time if we know only whitespace
has changed. But there is 1 case where cljfmt changes text: sort-ns-references.
To balance both multicursor nudging and sort-ns-references,
we format an ns form separately, then append the ordinarily-spaces-only
edits for the rest of the document. As a result, cursors located within
a (ns...) form might move to the end of the form, but other cursors
throughout the document should continue to float.